Fix damages screen crash issue, and clean the damagesForm
