+++ packages/react-native-views/src/components/CreateDamageForm/index.js
@@ -1,46 +1,19 @@
-import React, { useState, useCallback, useMemo, useEffect } from 'react';
+import React, { useMemo, useEffect } from 'react';
 import PropTypes from 'prop-types';
-import { ScrollView, StyleSheet, Text } from 'react-native';
-import { IconButton, DataTable, Button, withTheme, List } from 'react-native-paper';
-import { noop, startCase, snakeCase } from 'lodash';
+// import { View } from 'react-native';
+import { IconButton, withTheme, Provider as PaperProvider } from 'react-native-paper';
+import { noop, snakeCase } from 'lodash';
 
-import { utils } from '@monkvision/react-native';
-import Drawer from '../Drawer';
 import ImageViewer from '../ImageViewer';
-
 import useOrientation from '../../hooks/useOrientation';
 import useToggle from '../../hooks/useToggle';
 
 import damageMetadataList from './metadataList';
 
-import DamagePicturesPreview from './DamagePicturesPreview';
+import DamagesForm from './DamageForm';
 import CameraSimpleViewModal from './CameraSimpleViewModal/index';
-import DamageRow from './DamageRow';
 import DamagePicker from './DamagePicker';
-
-const { spacing } = utils.styles;
-const styles = StyleSheet.create({
-  cardTitle: {
-    fontSize: 16,
-  },
-  cardActions: {
-    display: 'flex',
-    justifyContent: 'center',
-    flexWrap: 'wrap',
-  },
-  alignLeft: { justifyContent: 'flex-end' },
-  buttonLabel: { color: '#FFFFFF' },
-  validationButton: {
-    margin: spacing(1),
-    flex: 1,
-    width: 120 },
-  clearButton: { alignSelf: 'flex-end' },
-  formWarningText: {
-    marginTop: spacing(2),
-    textAlign: 'center',
-    color: 'gray',
-  },
-});
+import useDamagesForm from './useDamagesForm';
 
 function CreateDamageForm({
   theme,
@@ -58,51 +31,21 @@ function CreateDamageForm({
 }) {
   const { colors } = theme;
   const [isCameraViewOpen, openCameraView, closeCameraView] = useToggle();
-  const [isSelectDialogOpen, openSelectDialog, closeSelectDialog] = useToggle();
-  const [isPreviewDialogOpen, openPreviewDialog, closePreviewDialog] = useToggle();
-
   const [damagePictures, setDamagePictures] = damagePicturesState;
-  const [previewImage, setPreviewImage] = useState({});
-  const [selectField, setSelectField] = useState(null);
 
-  const handleOpenPreviewDialog = useCallback((image) => {
-    setPreviewImage(image);
-    openPreviewDialog();
-  }, [openPreviewDialog]);
-
-  const handleOpenSelectDialog = useCallback((field) => {
-    setSelectField(field);
-    openSelectDialog();
-  }, [openSelectDialog]);
-
-  const handleDismissSelectDialog = useCallback(() => {
-    setSelectField(null);
-    closeSelectDialog();
-  }, [closeSelectDialog]);
-
-  const handleUpdateDamageMetaData = useCallback((metaData) => {
-    onChangeCurrentDamage(metaData);
-    handleDismissSelectDialog();
-  }, [handleDismissSelectDialog, onChangeCurrentDamage]);
-
-  const handleRemovePicture = useCallback(() => {
-    // Remove taken picture
-    setDamagePictures((prev) => prev.filter((_, i) => i !== previewImage.index));
-    closePreviewDialog();
-    setPreviewImage({});
-  }, [setDamagePictures, closePreviewDialog, previewImage.index]);
-
-  const handleClearDamagePictures = useCallback(() => setDamagePictures([]), [setDamagePictures]);
-
-  const wrapTitles = useMemo(() => {
-    if (currentDamage.damage_type) { return startCase(currentDamage.damage_type); }
-    return 'a damage';
-  }, [currentDamage.damage_type]);
-
-  const wrapSubtitles = useMemo(() => {
-    if (currentDamage.part_type) { return `On the ${startCase(currentDamage.part_type)} part`; }
-    return 'On a part';
-  }, [currentDamage.part_type]);
+  const {
+    isSelectDialogOpen,
+    isPreviewDialogOpen,
+    selectField,
+    handleOpenPreviewDialog,
+    handleOpenSelectDialog,
+    handleUpdateDamageMetaData,
+    handleRemovePicture,
+    handleClearDamagePictures,
+    previewImage,
+    closePreviewDialog,
+    handleDismissSelectDialog,
+  } = useDamagesForm({ onChangeCurrentDamage, setDamagePictures });
 
   const damagePicturesViewer = useMemo(() => ({
     isOpen: isPreviewDialogOpen,
@@ -121,122 +64,52 @@ function CreateDamageForm({
     if (isCameraViewOpen) { onCameraOpen(); } else { onCameraClose(); }
   }, [isCameraViewOpen, onCameraClose, onCameraOpen]);
 
-  // check if the form is filled with data and not yet submitted (dirty)
-  const isDirty = useMemo(() => Object.values(currentDamage).some(Boolean),
-    [currentDamage]);
-
   // camera view
   if (isCameraViewOpen) {
     return (
-      <CameraSimpleViewModal
-        theme={theme}
-        setDamagePictures={setDamagePictures}
-        closeCameraView={closeCameraView}
-        openPreviewDialog={handleOpenPreviewDialog}
-        damagePictures={damagePictures}
-        {...damagePicturesViewer}
-      />
+      <PaperProvider theme={theme}>
+        <CameraSimpleViewModal
+          theme={theme}
+          setDamagePictures={setDamagePictures}
+          closeCameraView={closeCameraView}
+          openPreviewDialog={handleOpenPreviewDialog}
+          damagePictures={damagePictures}
+          {...damagePicturesViewer}
+        />
+      </PaperProvider>
     );
   }
   // The drawer doesn't support orientation change (based on animated)
   // in this case we force the drawer to be closed on landscape orientation
   if (orientation !== 1 && !orientationIsNotSupported) { return null; }
   return (
-    <>
-      <Drawer
-        isOpen={isOpen}
-        handleClose={() => !isDirty && onClose()}
-        onClose={handleClearDamagePictures}
-      >
-        <ScrollView style={{ height: Drawer.CONTENT_HEIGHT }}>
-          <Drawer.Title
-            title={`Add photos for ${wrapTitles}`}
-            subtitle={wrapSubtitles}
-            titleStyle={styles.cardTitle}
-            left={(props) => <List.Icon {...props} icon="shape-square-plus" />}
-          />
-          <DamagePicturesPreview
-            damagePictures={damagePictures}
-            handleOpenPreviewDialog={handleOpenPreviewDialog}
-          />
-          <Drawer.Content>
-            <Button
-              accessibilityLabel="Reset form data"
-              mode="outlined"
-              icon="eraser-variant"
-              color={colors.primary}
-              style={styles.clearButton}
-              onPress={() => { onReset(); handleClearDamagePictures(); }}
-            >
-              Reset
-            </Button>
-            <DataTable>
-              <DataTable.Header>
-                <DataTable.Title>Add damage Metadata</DataTable.Title>
-                <DataTable.Title style={styles.alignLeft}>Value</DataTable.Title>
-              </DataTable.Header>
-              <DamageRow
-                title="Part type"
-                value={currentDamage.part_type}
-                key="metadata-partType"
-                onPress={() => handleOpenSelectDialog('part_type')}
-              />
-              <DamageRow
-                title="Damage type"
-                value={currentDamage.damage_type}
-                key="metadata-damageType"
-                onPress={() => handleOpenSelectDialog('damage_type')}
-              />
-              <DamageRow
-                title="Severity"
-                value={currentDamage.severity}
-                key="metadata-severity"
-                onPress={() => handleOpenSelectDialog('severity')}
-                disabled
-              />
-            </DataTable>
-            {isDirty ? (
-              <Text style={styles.formWarningText}>
-                In order to close the form please hit submit or clear your data.
-              </Text>
-            ) : null}
-          </Drawer.Content>
-          <Drawer.Actions style={styles.cardActions}>
-            <Button
-              accessibilityLabel="Add damage"
-              labelStyle={styles.buttonLabel}
-              onPress={onSubmit}
-              mode="contained"
-              style={styles.validationButton}
-              icon="shape-square-plus"
-              disabled={!isDamageValid}
-            >
-              Add damage
-            </Button>
-            <Button
-              accessibilityLabel="Add pictures"
-              onPress={openCameraView}
-              mode="outlined"
-              color={colors.primary}
-              style={[styles.validationButton, { borderColor: colors.primary }]}
-              icon="camera-plus"
-            >
-              Add pictures
-            </Button>
-          </Drawer.Actions>
-        </ScrollView>
-      </Drawer>
+    <PaperProvider theme={theme}>
+      <>
+        <DamagesForm
+          isOpen={isOpen}
+          onClose={onClose}
+          currentDamage={currentDamage}
+          isDamageValid={isDamageValid}
+          onSubmit={onSubmit}
+          onReset={onReset}
+          handleClearDamagePictures={handleClearDamagePictures}
+          damagePictures={damagePictures}
+          handleOpenPreviewDialog={handleOpenPreviewDialog}
+          handleOpenSelectDialog={handleOpenSelectDialog}
+          openCameraView={openCameraView}
+        />
 
-      {/* damage picker */}
-      <DamagePicker
-        visible={Boolean(isSelectDialogOpen && selectField)}
-        selectedValue={snakeCase(currentDamage[selectField]) || ''}
-        onValueChange={(value) => handleUpdateDamageMetaData({ [selectField]: value })}
-        data={damageMetadataList[selectField] || []}
-        onClose={handleDismissSelectDialog}
-      />
-      <ImageViewer {...damagePicturesViewer} />
-    </>
+        {/* damage picker */}
+        <DamagePicker
+          visible={Boolean(isSelectDialogOpen && selectField)}
+          selectedValue={snakeCase(currentDamage[selectField]) || ''}
+          onValueChange={(value) => handleUpdateDamageMetaData({ [selectField]: value })}
+          data={damageMetadataList[selectField] || []}
+          onClose={handleDismissSelectDialog}
+        />
+        <ImageViewer {...damagePicturesViewer} />
+      </>
+    </PaperProvider>
   );
 }
 CreateDamageForm.propTypes = {
