+++ packages/react-native-views/src/components/Damages/index.js
@@ -1,9 +1,10 @@
-import noop from 'lodash.noop';
-import PropTypes from 'prop-types';
+/* eslint-disable react/prop-types */
+// import noop from 'lodash.noop';
+// import PropTypes from 'prop-types';
 import React, { useImperativeHandle, forwardRef } from 'react';
-import { SafeAreaView } from 'react-native';
+import { View } from 'react-native';
 import { ActivityIndicatorView } from '@monkvision/react-native-views';
-
+import { Provider as PaperProvider } from 'react-native-paper';
 import useComputedParts from './useComputedParts';
 import useLayout from './useLayout';
 import usePartDamages from './usePartDamages';
@@ -14,6 +15,7 @@ import ValidateDialog from './ValidateDialog';
 import styles from './styles';
 
 function Damages({
+  theme,
   inspection,
   isLoading,
   isValidating,
@@ -44,64 +46,66 @@ function Damages({
   if (partsWithDamages.length === 0) { return null; }
 
   return (
-    <SafeAreaView style={styles.root}>
-      {isLoading ? <ActivityIndicatorView light /> : (
-        <>
-          <ValidateDialog
-            isDialogOpen={isDialogOpen}
-            handleDismissDialog={handleDismissDialog}
-            onValidate={onValidate}
-            isValidating={isValidating}
-            inspectionId={inspection.id}
-          />
-          <Navigation
-            partsWithDamages={partsWithDamages}
-            computedParts={computedParts}
-            damagedPartsCount={partsWithDamages.length}
-            handleOpenDialog={handleOpenDialog}
-            isValidated={isValidated}
-            onDeleteDamage={onDeleteDamage}
-            onSelectDamage={onSelectDamage}
-            isDeleting={isDeleting}
-            isVehiclePressAble={isVehiclePressAble && !isValidated}
-            onPressPart={onPressPart}
-            selectedId={selectedId}
-          />
-        </>
-      )}
-    </SafeAreaView>
+    <View style={styles.root}>
+      <PaperProvider theme={theme}>
+        {isLoading ? <ActivityIndicatorView light /> : (
+          <>
+            <ValidateDialog
+              isDialogOpen={isDialogOpen}
+              handleDismissDialog={handleDismissDialog}
+              onValidate={onValidate}
+              isValidating={isValidating}
+              inspectionId={inspection.id}
+            />
+            <Navigation
+              partsWithDamages={partsWithDamages}
+              computedParts={computedParts}
+              damagedPartsCount={partsWithDamages.length}
+              handleOpenDialog={handleOpenDialog}
+              isValidated={isValidated}
+              onDeleteDamage={onDeleteDamage}
+              onSelectDamage={onSelectDamage}
+              isDeleting={isDeleting}
+              isVehiclePressAble={isVehiclePressAble && !isValidated}
+              onPressPart={onPressPart}
+              selectedId={selectedId}
+            />
+          </>
+        )}
+      </PaperProvider>
+    </View>
   );
 }
 
-Damages.propTypes = {
-  inspection: PropTypes.shape({
-    // eslint-disable-next-line react/forbid-prop-types
-    damages: PropTypes.any.isRequired,
-    id: PropTypes.string.isRequired,
-    // eslint-disable-next-line react/forbid-prop-types
-    parts: PropTypes.any.isRequired,
-  }),
-  isDeleting: PropTypes.bool,
-  isLoading: PropTypes.bool,
-  isValidating: PropTypes.bool,
-  isVehiclePressAble: PropTypes.bool,
-  onDeleteDamage: PropTypes.func,
-  onPressPart: PropTypes.func,
-  onSelectDamage: PropTypes.func,
-  onValidate: PropTypes.func,
-  selectedId: PropTypes.string,
-};
+// Damages.propTypes = {
+//   inspection: PropTypes.shape({
+//     // eslint-disable-next-line react/forbid-prop-types
+//     damages: PropTypes.any.isRequired,
+//     id: PropTypes.string.isRequired,
+//     // eslint-disable-next-line react/forbid-prop-types
+//     parts: PropTypes.any.isRequired,
+//   }),
+//   isDeleting: PropTypes.bool,
+//   isLoading: PropTypes.bool,
+//   isValidating: PropTypes.bool,
+//   isVehiclePressAble: PropTypes.bool,
+//   onDeleteDamage: PropTypes.func,
+//   onPressPart: PropTypes.func,
+//   onSelectDamage: PropTypes.func,
+//   onValidate: PropTypes.func,
+//   selectedId: PropTypes.string,
+// };
 
-Damages.defaultProps = {
-  inspection: {},
-  isLoading: false,
-  isValidating: false,
-  isVehiclePressAble: false,
-  isDeleting: false,
-  onDeleteDamage: noop,
-  onSelectDamage: noop,
-  onPressPart: noop,
-  onValidate: noop,
-  selectedId: null,
-};
+// Damages.defaultProps = {
+//   inspection: {},
+//   isLoading: false,
+//   isValidating: false,
+//   isVehiclePressAble: false,
+//   isDeleting: false,
+//   onDeleteDamage: noop,
+//   onSelectDamage: noop,
+//   onPressPart: noop,
+//   onValidate: noop,
+//   selectedId: null,
+// };
 export default forwardRef(Damages);
