+++ packages/react-native-views/src/components/Damages/styles.js
@@ -6,7 +6,6 @@ export default StyleSheet.create({
     display: 'flex',
     width: '100%',
     height: '100%',
-    flex: 1,
   },
   scene: {
     display: 'flex',
